// src/app/models/customer-support.model.ts
export interface CustomerSupportModel {
    action:                     string;
    affectedProduct:            string;
    affectedProductId:          string;
    agreement:                  string;
    agreementEndDate:           string;
    agreementStartDate:         string;
    agreementType:              string;
    agreementValid:             string;
    applicationID:              string;
    callType:                   string;
    callingList:                string;
    campaignName:               string;
    caseId:                     string;
    caseType:                   string;
    channel:                    string;
    closingThresholdInMinutes:  string;
    component:                  string;
    componentId:                string;
    contactInfo:                string;
    contactOption:              string;
    createDate:                 string;
    customerCompanyName:        string;
    customerId:                 string;
    daysForCalls:               string;
    daysForCases:               string;
    daysForRepairs:             string;
    daysForSwitcher:            string;
    dialFromTime:               string;
    dialScheduledTime:          string;
    dialTillTime:               string;
    dsIdCountry:                string;
    dsIdLang:                   string;
    eligibleProduct:            string;
    eligibleProductId:          string;
    exceptionDescription:       string;
    exceptionTypeApplied:       string;
    firstName:                  string;
    geo:                        string;
    ipAddress:                  string;
    issue:                      string;
    issueId:                    string;
    language:                   string;
    lastName:                   string;
    maxRemoteAdvisor:           string;
    memberName:                 string;
    multipleCallsLimit:         string;
    multipleCasesLimit:         string;
    multipleRepairsLimit:       string;
    newCustomerIndicator:       string;
    numberOfCalls:              string;
    numberOfCases:              string;
    numberOfRepairs:            string;
    overrideIndicator:          string;
    platform:                   string;
    productGroup:               string;
    productGroupId:             string;
    productType:                string;
    purchaseAgreementIndicator: string;
    purchaseDate:               string;
    region:                     string;
    regionTarget:               string;
    riskFactor:                 string;
    safetyIssueIndicator:       string;
    scheduleNewContactInfo:     string;
    serialNumber:               string;
    skipCase:                   string;
    skipCaseReason:             string;
    smsCountryCode:             string;
    smsEncryptedCasLink:        string;
    smsLocale:                  string;
    smsMobileNumber:            string;
    smsReminderOptIn:           string;
    smsSurveyOptIn:             string;
    storeId:                    string;
    supportEntitlement:         string;
    timeZoneDatabaseId:         string;
    transactionId:              string;
    virtualAdvisorGroup:        string;
    virtualQueueGroup:          string;
    server:                     string;
}